Finally found the bug, will fix it tomorrow

+++ remake_printf/ft_call_pointer.c
@@ -1,6 +1,6 @@
 #include "libftprintf.h"
 
-int		ft_call_pointer(unsigned pointer, T_LIST *trail, char **print)
+int		ft_call_pointer(unsigned long pointer, T_LIST *trail, char **print)
 {
 	char    *result;
 	int		state_value;
